requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Data sources
yfinance>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sodapy>=1.2.0  # CFTC Socrata API client

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
streamlit>=1.28.0  # Web interface

# Excel export
openpyxl>=3.1.0
XlsxWriter>=3.1.0

# Configuration
pyyaml>=6.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.0.0
black>=23.0.0
